Menus Update
Settings:
Save setting right on value change and not by button.
Add defualt button to rest the settings

Stats:
'Reset' button now reset the player stats also

Upgrades:
Working upgrades menu

+++ Assets/Scripts/Stats/GameStats.cs
@@ -1,10 +1,11 @@
 using System;
+using UnityEngine;
 
 [Serializable]
 public class GameStats
 {
     public int RecordScore;
-    public int TotalCoins;
+    public float TotalCoins;
 
     public GameStats() 
     {
